Move TextEditingDelta to its own file
