+++ packages/flutter/lib/services.dart
@@ -42,6 +42,7 @@ export 'src/services/system_chrome.dart';
 export 'src/services/system_navigator.dart';
 export 'src/services/system_sound.dart';
 export 'src/services/text_editing.dart';
+export 'src/services/text_editing_delta.dart';
 export 'src/services/text_formatter.dart';
 export 'src/services/text_input.dart';
 export 'src/services/text_layout_metrics.dart';
